+++ piccolo/query/methods/objects.py
@@ -73,8 +73,25 @@ class GetOrCreate(t.Generic[TableInstance]):
 
         await instance.save().run()
 
-        instance._was_created = True
+        # If the user wants us to prefetch related objects, for example:
+        #
+        # await Band.objects(Band.manager).get_or_create(
+        #   (Band.name == 'Pythonistas') & (Band.manager == 1)
+        # )
+        #
+        # Then we need to fetch the related objects.
+        # See https://github.com/piccolo-orm/piccolo/issues/597
+        prefetch = self.query.prefetch_delegate.fk_columns
+        if prefetch:
+            table = instance.__class__
+            primary_key = table._meta.primary_key
+            instance = (
+                await table.objects(*prefetch)
+                .get(primary_key == getattr(instance, primary_key._meta.name))
+                .run()
+            )
 
+        instance._was_created = True
         return instance
 
     def __await__(self) -> t.Generator[None, None, TableInstance]:
@@ -258,10 +275,15 @@ class Objects(Query, t.Generic[TableInstance]):
         return self
 
     async def batch(
-        self: Self, batch_size: t.Optional[int] = None, **kwargs
+        self: Self,
+        batch_size: t.Optional[int] = None,
+        node: t.Optional[str] = None,
+        **kwargs
     ) -> Batch:
         if batch_size:
             kwargs.update(batch_size=batch_size)
+        if node:
+            kwargs.update(node=node)
         return await self.table._meta.db.batch(self, **kwargs)
 
     async def response_handler(self, response):