+++ tests/apps/migrations/commands/test_forwards_backwards.py
@@ -24,7 +24,6 @@ from tests.example_apps.music.tables import (
 if t.TYPE_CHECKING:  # pragma: no cover
     from piccolo.table import Table
 
-
 TABLE_CLASSES: t.List[t.Type[Table]] = [
     Manager,
     Band,
@@ -53,6 +52,7 @@ class TestForwardsBackwards(TestCase):
             # Check the tables exist
             for table_class in TABLE_CLASSES:
                 self.assertTrue(table_class.table_exists().run_sync())
+            self.assertNotEqual(Migration.count().run_sync(), 0)
 
             run_sync(
                 backwards(
@@ -63,19 +63,33 @@ class TestForwardsBackwards(TestCase):
             # Check the tables don't exist
             for table_class in TABLE_CLASSES:
                 self.assertTrue(not table_class.table_exists().run_sync())
+            self.assertEqual(Migration.count().run_sync(), 0)
+            # Preview
+            run_sync(
+                forwards(app_name=app_name, migration_id="all", preview=True)
+            )
+            for table_class in TABLE_CLASSES:
+                self.assertTrue(not table_class.table_exists().run_sync())
+            self.assertEqual(Migration.count().run_sync(), 0)
 
     def test_forwards_backwards_single_migration(self):
         """
         Test running a single migrations forwards, then backwards.
         """
+        table_classes = [Band, Manager]
+
         for migration_id in ["1", "2020-12-17T18:44:30"]:
             run_sync(forwards(app_name="music", migration_id=migration_id))
-
-            table_classes = [Band, Manager]
 
             # Check the tables exist
             for table_class in table_classes:
                 self.assertTrue(table_class.table_exists().run_sync())
+
+            self.assertTrue(
+                Migration.exists()
+                .where(Migration.name == "2020-12-17T18:44:30")
+                .run_sync()
+            )
 
             run_sync(
                 backwards(
@@ -88,6 +102,25 @@ class TestForwardsBackwards(TestCase):
             # Check the tables don't exist
             for table_class in table_classes:
                 self.assertTrue(not table_class.table_exists().run_sync())
+            self.assertFalse(
+                Migration.exists()
+                .where(Migration.name == "2020-12-17T18:44:30")
+                .run_sync()
+            )
+
+            # Preview
+            run_sync(
+                forwards(
+                    app_name="music", migration_id=migration_id, preview=True
+                )
+            )
+            for table_class in table_classes:
+                self.assertTrue(not table_class.table_exists().run_sync())
+            self.assertFalse(
+                Migration.exists()
+                .where(Migration.name == "2020-12-17T18:44:30")
+                .run_sync()
+            )
 
     @patch("piccolo.apps.migrations.commands.forwards.print")
     def test_forwards_unknown_migration(self, print_: MagicMock):