+++ tests/table/test_insert.py
@@ -1,4 +1,6 @@
-from tests.base import DBTestCase
+import pytest
+
+from tests.base import DBTestCase, engine_version_lt, is_running_sqlite
 from tests.example_apps.music.tables import Band, Manager
 
 
@@ -42,3 +44,35 @@ class TestInsert(DBTestCase):
         names = [i["name"] for i in response]
 
         self.assertIn("{}", names)
+
+    @pytest.mark.skipif(
+        is_running_sqlite() and engine_version_lt(3.35),
+        reason="SQLite version not supported",
+    )
+    def test_insert_returning(self):
+        """
+        Make sure update works with the `returning` clause.
+        """
+        response = (
+            Manager.insert(Manager(name="Maz"))
+            .returning(Manager.name)
+            .run_sync()
+        )
+
+        self.assertListEqual(response, [{"name": "Maz"}])
+
+    @pytest.mark.skipif(
+        is_running_sqlite() and engine_version_lt(3.35),
+        reason="SQLite version not supported",
+    )
+    def test_insert_returning_alias(self):
+        """
+        Make sure update works with the `returning` clause.
+        """
+        response = (
+            Manager.insert(Manager(name="Maz"))
+            .returning(Manager.name.as_alias("manager_name"))
+            .run_sync()
+        )
+
+        self.assertListEqual(response, [{"manager_name": "Maz"}])